cont.

+++ js/main.js
@@ -41,7 +41,7 @@ app = new Vue({
         },
         change_bet: function () {
             play_se("se_click");
-            var new_bet = parseInt(prompt("赌多少？"));
+            var new_bet = parseFloat(prompt("赌多少？"));
             // Check new bet
             if (new_bet > 0) {
                 this.bet_input = new Number(new_bet).toFixed(4);
